\name{function}
\alias{function}
\title{A title for the function}
\description{A description for the function}
\usage{
function()}
\arguments{
  \item{argument}{A description}
}
\details{
Give details about the function. 
}
\value{
What is returned by the function?
}
\references{
Some references
}
\author{
List the authors here.
}
\examples{
Provide examples on how to run the function.
}